Revert "refactor(FormControl): update FormControl to use CSS Modules behind f…" (#5494)
This reverts commit 4776d13951bf72c2dfad427dcd07fadcce2c40b3.
