--- .changeset/gentle-stingrays-search.md
@@ -1,5 +1,0 @@
----
-"@primer/react": minor
----
-
-Update FormControl to use CSS Modules behind feature flag
