--- packages/react/src/FormControl/feature-flags.ts
@@ -1,1 +1,0 @@
-export const cssModulesFlag = 'primer_react_css_modules_team'
