+++ packages/react/src/__tests__/FormControl.test.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import {render} from '@testing-library/react'
+import {renderHook} from '@testing-library/react-hooks'
 import axe from 'axe-core'
 import {
   Autocomplete,
@@ -11,7 +12,8 @@ import {
   Textarea,
   TextInput,
   TextInputWithTokens,
-} from '../..'
+  useFormControlForwardedProps,
+} from '..'
 import {MarkGithubIcon} from '@primer/octicons-react'
 
 const LABEL_TEXT = 'Form control'
@@ -450,5 +452,55 @@ describe('FormControl', () => {
       const results = await axe.run(container)
       expect(results).toHaveNoViolations()
     })
+  })
+})
+
+describe('useFormControlForwardedProps', () => {
+  describe('when used outside FormControl', () => {
+    test('returns empty object when no props object passed', () => {
+      const result = renderHook(() => useFormControlForwardedProps({}))
+      expect(result.result.current).toEqual({})
+    })
+
+    test('returns passed props object instance when passed', () => {
+      const props = {id: 'test-id'}
+      const result = renderHook(() => useFormControlForwardedProps(props))
+      expect(result.result.current).toBe(props)
+    })
+  })
+
+  test('provides context value when no props object is passed', () => {
+    const id = 'test-id'
+
+    const {result} = renderHook(() => useFormControlForwardedProps({}), {
+      wrapper: ({children}: {children: React.ReactNode}) => (
+        <FormControl id={id} disabled required>
+          <FormControl.Label>Label</FormControl.Label>
+          {children}
+        </FormControl>
+      ),
+    })
+
+    expect(result.current.disabled).toBe(true)
+    expect(result.current.id).toBe(id)
+    expect(result.current.required).toBe(true)
+  })
+
+  test('merges with props object, overriding to prioritize props when conflicting', () => {
+    const props = {id: 'override-id', xyz: 'someValue'}
+
+    const {result} = renderHook(() => useFormControlForwardedProps(props), {
+      wrapper: ({children}: {children: React.ReactNode}) => (
+        <FormControl id="form-control-id" disabled>
+          <FormControl.Label>Label</FormControl.Label>
+          {children}
+        </FormControl>
+      ),
+    })
+
+    expect(result.current.disabled).toBe(true)
+    expect(result.current.id).toBe(props.id)
+    expect(result.current.required).toBeFalsy()
+    expect(result.current.xyz).toBe(props.xyz)
   })
 })
