+++ packages/react/src/internal/components/InputLabel.tsx
@@ -1,10 +1,7 @@
-import {clsx} from 'clsx'
 import React from 'react'
 import styled from 'styled-components'
+import {get} from '../../constants'
 import sx, {type SxProp} from '../../sx'
-import {cssModulesFlag} from '../../FormControl/feature-flags'
-import {useFeatureFlag} from '../../FeatureFlags'
-import classes from './InputLabel.module.css'
 
 type BaseProps = SxProp & {
   disabled?: boolean
@@ -42,7 +39,6 @@ function InputLabel({
   className,
   ...props
 }: Props) {
-  const enabled = useFeatureFlag(cssModulesFlag)
   return (
     <StyledLabel
       as={as}
@@ -50,18 +46,12 @@ function InputLabel({
       data-visually-hidden={visuallyHidden ? '' : undefined}
       htmlFor={htmlFor}
       id={id}
-      className={clsx(className, {
-        [classes.Label]: enabled,
-      })}
+      className={className}
       sx={sx}
       {...props}
     >
       {required || requiredText ? (
-        <StyledRequiredText
-          className={clsx({
-            [classes.RequiredText]: enabled,
-          })}
-        >
+        <StyledRequiredText>
           <span>{children}</span>
           <span aria-hidden={requiredIndicator ? undefined : true}>{requiredText ?? '*'}</span>
         </StyledRequiredText>
@@ -72,13 +62,18 @@ function InputLabel({
   )
 }
 
+const StyledRequiredText = styled.span`
+  display: flex;
+  column-gap: ${get('space.1')};
+`
+
 const StyledLabel = styled.label`
   align-self: flex-start;
   display: block;
   color: var(--fgColor-default);
   cursor: pointer;
   font-weight: 600;
-  font-size: var(--text-body-size-medium);
+  font-size: ${get('fontSizes.1')};
 
   &:where([data-control-disabled]) {
     color: var(--fgColor-muted);
@@ -99,11 +94,6 @@ const StyledLabel = styled.label`
   }
 
   ${sx}
-`
-
-const StyledRequiredText = styled.span`
-  display: flex;
-  column-gap: var(--base-size-4);
 `
 
 export {InputLabel}
