+++ packages/react/src/internal/components/InputValidation.tsx
@@ -1,15 +1,12 @@
 import type {IconProps} from '@primer/octicons-react'
 import {AlertFillIcon, CheckCircleFillIcon} from '@primer/octicons-react'
-import {clsx} from 'clsx'
 import React from 'react'
-import styled from 'styled-components'
 import Text from '../../Text'
-import sx from '../../sx'
 import type {SxProp} from '../../sx'
-import {cssModulesFlag} from '../../FormControl/feature-flags'
 import type {FormValidationStatus} from '../../utils/types/FormValidationStatus'
-import {useFeatureFlag} from '../../FeatureFlags'
-import classes from './InputValidation.module.css'
+import styled from 'styled-components'
+import {get} from '../../constants'
+import sx from '../../sx'
 
 type Props = {
   id: string
@@ -25,7 +22,6 @@ const validationIconMap: Record<
 }
 
 const InputValidation: React.FC<React.PropsWithChildren<Props>> = ({children, id, validationStatus, sx}) => {
-  const enabled = useFeatureFlag(cssModulesFlag)
   const IconComponent = validationStatus ? validationIconMap[validationStatus] : undefined
 
   // TODO: use `text-caption-lineHeight` token as a custom property when it's available
@@ -35,19 +31,10 @@ const InputValidation: React.FC<React.PropsWithChildren<Props>> = ({children, id
   const iconBoxMinHeight = iconSize * captionLineHeight
 
   return (
-    <StyledInputValidation
-      className={clsx({
-        [classes.InputValidation]: enabled,
-      })}
-      data-validation-status={validationStatus}
-      sx={sx}
-    >
+    <StyledInputValidation data-validation-status={validationStatus} sx={sx}>
       {IconComponent ? (
         <StyledValidationIcon
           aria-hidden="true"
-          className={clsx({
-            [classes.ValidationIcon]: enabled,
-          })}
           style={
             {
               '--inputValidation-iconSize': iconBoxMinHeight,
@@ -57,13 +44,7 @@ const InputValidation: React.FC<React.PropsWithChildren<Props>> = ({children, id
           <IconComponent size={iconSize} fill="currentColor" />
         </StyledValidationIcon>
       ) : null}
-      <StyledValidationText
-        id={id}
-        className={clsx({
-          [classes.ValidationText]: enabled,
-        })}
-        style={{'--inputValidation-lineHeight': captionLineHeight} as React.CSSProperties}
-      >
+      <StyledValidationText id={id} style={{'--inputValidation-lineHeight': captionLineHeight} as React.CSSProperties}>
         {children}
       </StyledValidationText>
     </StyledInputValidation>
@@ -73,7 +54,7 @@ const InputValidation: React.FC<React.PropsWithChildren<Props>> = ({children, id
 const StyledInputValidation = styled(Text)`
   color: var(--inputValidation-fgColor);
   display: flex;
-  font-size: var(--text-body-size-small);
+  font-size: ${get('fontSizes.0')};
   font-weight: 600;
 
   & :where(a) {
@@ -82,11 +63,11 @@ const StyledInputValidation = styled(Text)`
   }
 
   &:where([data-validation-status='success']) {
-    --inputValidation-fgColor: var(--fgColor-success);
+    --inputValidation-fgColor: ${get('colors.success.fg')};
   }
 
   &:where([data-validation-status='error']) {
-    --inputValidation-fgColor: var(--fgColor-danger);
+    --inputValidation-fgColor: ${get('colors.danger.fg')};
   }
 
   ${sx}
@@ -95,7 +76,7 @@ const StyledInputValidation = styled(Text)`
 const StyledValidationIcon = styled.span`
   align-items: center;
   display: flex;
-  margin-inline-end: var(--base-size-4);
+  margin-inline-end: ${get('space.1')};
   min-height: var(--inputValidation-iconSize);
 `
